+++ tests/Unit/Twig/NowoCkeditor5EditorTwigExtensionTest.php
@@ -12,41 +12,40 @@ use PHPUnit\Framework\TestCase;
  */
 final class NowoCkeditor5EditorTwigExtensionTest extends TestCase
 {
-    public function testGetFunctionsReturnsAssetPathFunction(): void
+    public function testGetFunctionsReturnsAssetHelpers(): void
     {
         $ext = new NowoCkeditor5EditorTwigExtension();
         $fns = $ext->getFunctions();
 
-        self::assertCount(1, $fns);
+        self::assertCount(2, $fns);
         self::assertSame('nowo_ckeditor5_editor_asset_path', $fns[0]->getName());
+        self::assertSame('nowo_ckeditor5_editor_asset_package', $fns[1]->getName());
     }
 
-    public function testAssetPathReturnsPathWithAssetDir(): void
+    public function testAssetPathReturnsRelativePath(): void
     {
         $ext = new NowoCkeditor5EditorTwigExtension();
 
-        self::assertSame('bundles/nowockeditor5editor/ckeditor5-editor.js', $ext->assetPath('ckeditor5-editor.js'));
-        self::assertSame('bundles/nowockeditor5editor/ckeditor5-editor.js', $ext->assetPath('/ckeditor5-editor.js'));
+        self::assertSame('ckeditor5-editor.js', $ext->assetPath('ckeditor5-editor.js'));
+        self::assertSame('ckeditor5-editor.js', $ext->assetPath('/ckeditor5-editor.js'));
     }
 
     public function testAssetPathRejectsPathTraversal(): void
     {
-        $ext     = new NowoCkeditor5EditorTwigExtension();
-        $default = 'bundles/' . NowoCkeditor5EditorTwigExtension::ASSET_DIR . '/ckeditor5-editor.js';
-        self::assertSame($default, $ext->assetPath('../other/file.js'));
+        $ext = new NowoCkeditor5EditorTwigExtension();
+        self::assertSame('ckeditor5-editor.js', $ext->assetPath('../other/file.js'));
     }
 
     public function testAssetPathRejectsInvalidCharacters(): void
     {
-        $ext     = new NowoCkeditor5EditorTwigExtension();
-        $default = 'bundles/' . NowoCkeditor5EditorTwigExtension::ASSET_DIR . '/ckeditor5-editor.js';
-        self::assertSame($default, $ext->assetPath('bad<script>.js'));
-        self::assertSame($default, $ext->assetPath(''));
+        $ext = new NowoCkeditor5EditorTwigExtension();
+        self::assertSame('ckeditor5-editor.js', $ext->assetPath('bad<script>.js'));
+        self::assertSame('ckeditor5-editor.js', $ext->assetPath(''));
     }
 
     public function testAssetPathAllowsSubpath(): void
     {
         $ext = new NowoCkeditor5EditorTwigExtension();
-        self::assertSame('bundles/nowockeditor5editor/css/theme.css', $ext->assetPath('css/theme.css'));
+        self::assertSame('css/theme.css', $ext->assetPath('css/theme.css'));
     }
 }